pyproject.toml
[tool.black]
line-length = 88
target-version = ['py311']
exclude = 'migrations'

[tool.poetry]
name = "proposal-extraction-benchmarks"
description = "Benchmark PDF extractors for insurance proposal data extraction"
version = "0.1.0"
package-mode = false

[tool.poetry.dependencies]
black = "22.12.0"
python = "^3.11"
openai = "^1.0"
python-dotenv = "^1.0"
pymupdf = "^1.24"
pdfplumber = "^0.11"
pypdfium2 = "^4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"